[Refactor] gRPC 중복 호출 오류 수정
