+++ fridaymarket_resource/src/main/java/com/smile/fridaymarket_resource/grpc/service/GrpcAuthService.java
@@ -3,6 +3,7 @@ package com.smile.fridaymarket_resource.grpc.service;
 import com.smile.fridaymarket_auth.grpc.AuthTokenProto;
 import com.smile.fridaymarket_auth.grpc.AuthTokenServiceGrpc;
 import com.smile.fridaymarket_resource.auth.UserResponse;
+import com.smile.fridaymarket_resource.grpc.interceptor.JwtClientInterceptor;
 import io.grpc.StatusRuntimeException;
 import lombok.extern.slf4j.Slf4j;
 import net.devh.boot.grpc.client.inject.GrpcClient;
@@ -16,20 +17,24 @@ public class GrpcAuthService {
     private AuthTokenServiceGrpc.AuthTokenServiceBlockingStub authStub;
 
     public UserResponse authToken(String accessToken) {
+
         try {
-            AuthTokenProto.AuthTokenResponse authTokenResponse = authStub.verifyToken(
+            // 인터셉터를 추가한 Stub 생성 (주입된 authStub을 사용)
+            AuthTokenServiceGrpc.AuthTokenServiceBlockingStub interceptedStub =
+                    authStub.withInterceptors(new JwtClientInterceptor(accessToken));
+
+            AuthTokenProto.AuthTokenResponse authTokenResponse = interceptedStub.verifyToken(
                     AuthTokenProto.AuthTokenRequest.newBuilder()
                             .setAccessToken(accessToken)
                             .build()
             );
+
             log.info("gRPC 통신 응답 username: {}", authTokenResponse.getUsername());
 
             return new UserResponse(true, authTokenResponse.getUserId(), authTokenResponse.getUsername());
         }
         catch (StatusRuntimeException e) {
-            // gRPC 호출 실패 시 오류 메시지 기록
             log.info("gRPC 호출 실패 : {}", e.getStatus().getCode().name());
-            // 인증 실패 응답 반환
             return new UserResponse(false, "Unknown", "Unknown");
         }
     }
